Add support for intermediate stage 1.5
Introduce an intermediate "stage 1.5" (The Bridge) by changing stage typing to float across env, reward, and state setter APIs. build_pinch_reward now accepts 1.5 (treated like stage 1 for goal logic) and the error message lists 1.5. PinchGoldenSeedSetter gains configuration for stage 1.5 (reduced noise and smaller rewind) and applies a dynamic rewind for that stage. test_stage2_spawns.py now accepts a --stage float argument and prints the active stage when showing spawns.

+++ src/envs/pinch.py
@@ -20,7 +20,7 @@ from rewards.pinch_reward import build_pinch_reward, build_golden_seed_reward
 from state_setters.pinch_golden_seed_setter import PinchGoldenSeedSetter
 
 
-def build_env(render: bool = False, tick_skip: int = 8, stage: int = 1, difficulty_level: int = 1):
+def build_env(render: bool = False, tick_skip: int = 8, stage: float = 1.0, difficulty_level: int = 1):
     """
     Build the pinch specialist environment.
